Refactored additions to commandbox as modules
