+++ src/main/java/seedu/address/logic/textFieldModules/FzfModule.java
@@ -1,4 +1,4 @@
-package seedu.address.ui;
+package seedu.address.logic.textFieldModules;
 
 import java.util.List;
 import java.util.function.Supplier;
